postgres_output: write to specific columns
This is now required in PostgresOutput. Safer to default to strictness.
Since the previous way you wrote values in order to columns in a table,
that implicitly showed knowledge of the schema. Might as well enforce it
and thereby be robust to adding/removing columns not used in this output.