lua/decoders: kvmeta- add more meta beyond `routes`
Previously, we just wrote `routes`. Now, we also write other kvmeta to
the message, like `kv_language` or `kv_version`.

This data gets flattened and written alongside the route. So we should
ensure these keys don't conflict with keys in the route. for `kv_<x>`
this is easy, since they're prepended by `kv` it's highly unlikely to
conflict.